pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "llm-proxy"
version = "0.1.0"
description = "Multi-protocol LLM reverse proxy with dual ingress (Anthropic Messages + OpenAI Chat Completions) and multi-provider upstream routing."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic>=2.11.0,<3.0.0",
  "pyyaml>=6.0.2,<7.0.0",
  "uvicorn>=0.35.0,<1.0.0",
]

[project.scripts]
llm-proxy = "llm_proxy.__main__:main"

[project.optional-dependencies]
dev = [
  "pytest>=8.4.0,<9.0.0",
  "pytest-asyncio>=1.1.0,<2.0.0",
  "respx>=0.22.0,<1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["llm_proxy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
